Fix all the specs

+++ app/graphql/resolvers/visitors/pages_per_session.rb
@@ -3,7 +3,7 @@
 module Resolvers
   module Visitors
     class PagesPerSession < Resolvers::Base
-      type Integer, null: false
+      type Float, null: false
 
       def resolve
         sql = <<-SQL
